[kbss-cvut/termit-ui#303] Pull types to AbstractTerm so that they are available in TermDto.
This ensures snapshot type is available in DTO for frontend navigation.
